Enhance AIRecommendationScreen and DashboardScreen to support detailed predictions. Updated data handling for AI predictions, ensuring robust calculations for remaining time. Changed ai_prediction type in the database schema from integer to numeric for improved accuracy.

+++ project/supabase/migrations/20250613192929_broad_cliff.sql
@@ -95,7 +95,7 @@ CREATE TABLE IF NOT EXISTS flower_batches (
   floral_food_used boolean NOT NULL DEFAULT false,
   vase_cleanliness text NOT NULL DEFAULT 'Clean',
   dynamic_spoilage_date timestamptz NOT NULL,
-  ai_prediction integer,
+  ai_prediction numeric,
   ai_confidence numeric,
   ai_reasoning text,
   ai_recommendations text[],
